Fix bug: Changing manually the tick of keyframes causes to ignore decimals
Refactor keyframe handling to support float precision

- Updated keyframe-related methods and classes to use float instead of int for tick values, enhancing precision in keyframe management.
- Improved UI interactions by ensuring that updates do not interfere with user input during editing.
- Added checks to prevent unnecessary updates when input fields are focused, improving user experience during keyframe editing.

+++ src/client/java/mchorse/bbs_mod/ui/film/replays/UIReplaysEditor.java
@@ -1346,7 +1346,7 @@ public class UIReplaysEditor extends UIElement implements GizmoSurface
         }
     }
 
-    public void insertKeyframeColumn(int tick)
+    public void insertKeyframeColumn(float tick)
     {
         Replay replay = this.getReplay();
         IEntity entity = this.filmPanel.getController().getCurrentEntity();
@@ -1364,7 +1364,7 @@ public class UIReplaysEditor extends UIElement implements GizmoSurface
         });
     }
 
-    public void insertKeyframeColumnInterpolated(int tick)
+    public void insertKeyframeColumnInterpolated(float tick)
     {
         Replay replay = this.getReplay();
